+++ agent.py
@@ -19,6 +19,8 @@ import os
 import signal
 import sys
 import time
+import supervisor.xmlrpc
+import xmlrpclib
 from copy import copy
 
 # For pickle & PID files, see issue 293
@@ -29,13 +31,19 @@ from checks.check_status import CollectorStatus
 from checks.collector import Collector
 from config import (
     get_config,
+    get_jmx_pipe_path,
     get_parsed_args,
     get_system_stats,
     load_check_directory,
-    load_check
+    load_check,
+    generate_jmx_configs,
+    _is_affirmative,
+    SD_PIPE_NAME
+
 )
 from daemon import AgentSupervisor, Daemon
 from emitter import http_emitter
+from utils.platform import Platform
 
 # utils
 from util import Watchdog
@@ -50,13 +58,18 @@ from utils.service_discovery.config_stores import get_config_store
 from utils.service_discovery.sd_backend import get_sd_backend
 
 # Constants
+from jmxfetch import JMX_CHECKS
 PID_NAME = "dd-agent"
 PID_DIR = None
 WATCHDOG_MULTIPLIER = 10
 RESTART_INTERVAL = 4 * 24 * 60 * 60  # Defaults to 4 days
-START_COMMANDS = ['start', 'restart', 'foreground']
-DD_AGENT_COMMANDS = ['check', 'flare', 'jmx']
 
+JMX_SUPERVISOR_ENTRY = 'datadog-agent:jmxfetch'
+JMX_GRACE_SECS = 2
+SERVICE_DISCOVERY_PREFIX = 'SD-'
+SD_CONFIG_SEP = "#### SERVICE-DISCOVERY ####\n"
+
+DEFAULT_SUPERVISOR_SOCKET = '/opt/datadog-agent/run/datadog-supervisor.sock'
 DEFAULT_COLLECTOR_PROFILE_INTERVAL = 20
 
 # Globals
@@ -81,6 +94,9 @@ class Agent(Daemon):
         # this flag can be set to True, False, or a list of checks (for partial reload)
         self.reload_configs_flag = False
         self.sd_backend = None
+        self.supervisor_proxy = None
+        self.sd_pipe = None
+
 
     def _handle_sigterm(self, signum, frame):
         """Handles SIGTERM and SIGINT, which gracefully stops the agent."""
@@ -106,6 +122,7 @@ class Agent(Daemon):
            Can also reload only an explicit set of checks."""
         log.info("Attempting a configuration reload...")
         hostname = get_hostname(self._agentConfig)
+        jmx_sd_configs = None
 
         # if no check was given, reload them all
         if not checks_to_reload:
@@ -115,12 +132,22 @@ class Agent(Daemon):
                 check.stop()
 
             self._checksd = load_check_directory(self._agentConfig, hostname)
+            if self._jmx_service_discovery_enabled:
+                jmx_sd_configs = generate_jmx_configs(self._agentConfig, hostname)
         else:
             new_checksd = copy(self._checksd)
 
-            self.refresh_specific_checks(hostname, new_checksd, checks_to_reload)
+            jmx_checks = [check for check in checks_to_reload if check in JMX_CHECKS]
+            py_checks = set(checks_to_reload) - set(jmx_checks)
+            self.refresh_specific_checks(hostname, new_checksd, py_checks)
+            if self._jmx_service_discovery_enabled:
+                jmx_sd_configs = generate_jmx_configs(self._agentConfig, hostname, jmx_checks)
+
             # once the reload is done, replace existing checks with the new ones
             self._checksd = new_checksd
+
+        if jmx_sd_configs:
+            self._submit_jmx_service_discovery(jmx_sd_configs)
 
         # Logging
         num_checks = len(self._checksd['initialized_checks'])
@@ -218,8 +245,31 @@ class Agent(Daemon):
         if self._agentConfig.get('service_discovery'):
             self.sd_backend = get_sd_backend(self._agentConfig)
 
+        if _is_affirmative(self._agentConfig.get('sd_jmx_enable')):
+            pipe_path = get_jmx_pipe_path()
+            if Platform.is_windows():
+                pipe_name = pipe_path.format(pipename=SD_PIPE_NAME)
+            else:
+                pipe_name = os.path.join(pipe_path, SD_PIPE_NAME)
+
+            if os.access(pipe_path, os.W_OK):
+                if not os.path.exists(pipe_name):
+                    os.mkfifo(pipe_name)
+                self.sd_pipe = os.open(pipe_name, os.O_RDWR) # RW to avoid blocking (will only W)
+
+                # Initialize Supervisor proxy
+                self.supervisor_proxy = self._get_supervisor_socket(self._agentConfig)
+            else:
+                log.debug('Unable to create pipe in temporary directory. JMX service discovery disabled.')
+
         # Load the checks.d checks
         self._checksd = load_check_directory(self._agentConfig, hostname)
+
+        # Load JMX configs if available
+        if self._jmx_service_discovery_enabled:
+            jmx_sd_configs = generate_jmx_configs(self._agentConfig, hostname)
+            if jmx_sd_configs:
+                self._submit_jmx_service_discovery(jmx_sd_configs)
 
         # Initialize the Collector
         self.collector = Collector(self._agentConfig, emitters, systemStats, hostname)
@@ -241,13 +291,15 @@ class Agent(Daemon):
         self.restart_interval = int(self._agentConfig.get('restart_interval', RESTART_INTERVAL))
         self.agent_start = time.time()
 
+        self.allow_profiling = self._agentConfig.get('allow_profiling', True)
+
         profiled = False
         collector_profiled_runs = 0
 
         # Run the main loop.
         while self.run_forever:
             # Setup profiling if necessary
-            if self.in_developer_mode and not profiled:
+            if self.allow_profiling and self.in_developer_mode and not profiled:
                 try:
                     profiler = AgentProfiler()
                     profiler.enable_profiling()
@@ -344,6 +396,52 @@ class Agent(Daemon):
                 log.info('Not running on EC2, using hostname to identify this server')
         return agentConfig
 
+    def _get_supervisor_socket(self, agentConfig):
+        if Platform.is_windows():
+            return None
+
+        sockfile = agentConfig.get('supervisor_socket', DEFAULT_SUPERVISOR_SOCKET)
+        supervisor_proxy = xmlrpclib.ServerProxy(
+            'http://127.0.0.1',
+            transport=supervisor.xmlrpc.SupervisorTransport(
+                None, None, serverurl="unix://{socket}".format(socket=sockfile))
+        )
+
+        return supervisor_proxy
+
+    @property
+    def _jmx_service_discovery_enabled(self):
+        return self.sd_pipe is not None
+
+    def _submit_jmx_service_discovery(self, jmx_sd_configs):
+
+        if not jmx_sd_configs or not self.sd_pipe:
+            return
+
+        if self.supervisor_proxy is not None:
+            jmx_state = self.supervisor_proxy.supervisor.getProcessInfo(JMX_SUPERVISOR_ENTRY)
+            log.debug("Current JMX check state: %s", jmx_state['statename'])
+            # restart jmx if stopped
+            if jmx_state['statename'] in ['STOPPED', 'EXITED', 'FATAL'] and self._agentConfig.get('sd_jmx_enable'):
+                self.supervisor_proxy.supervisor.startProcess(JMX_SUPERVISOR_ENTRY)
+                time.sleep(JMX_GRACE_SECS)
+        else:
+            log.debug("Unable to automatically start jmxfetch on Windows via supervisor.")
+
+        buffer = ""
+        for name, yaml in jmx_sd_configs.iteritems():
+            try:
+                buffer += SD_CONFIG_SEP
+                buffer += "# {}\n".format(name)
+                buffer += yaml
+            except Exception as e:
+                log.exception("unable to submit YAML via RPC: %s", e)
+            else:
+                log.info("JMX SD Config via named pip %s successfully.", name)
+
+        if buffer:
+            os.write(self.sd_pipe, buffer)
+
     def _should_restart(self):
         if time.time() - self.agent_start > self.restart_interval:
             return True
@@ -362,6 +460,7 @@ def main():
     autorestart = agentConfig.get('autorestart', False)
     hostname = get_hostname(agentConfig)
     in_developer_mode = agentConfig.get('developer_mode')
+
     COMMANDS_AGENT = [
         'start',
         'stop',
@@ -389,17 +488,13 @@ def main():
         sys.stderr.write("Unknown command: %s\n" % command)
         return 3
 
-    # Deprecation notice
-    if command not in DD_AGENT_COMMANDS:
-        # Will become an error message and exit after deprecation period
-        from utils.deprecations import deprecate_old_command_line_tools
-        deprecate_old_command_line_tools()
+    # TODO: actually kill the start/stop/restart/status command for 5.11
+    if command in ['start', 'stop', 'restart', 'status'] and not in_developer_mode:
+        logging.error('Please use supervisor to manage the agent')
+        return 1
 
     if command in COMMANDS_AGENT:
         agent = Agent(PidFile(PID_NAME, PID_DIR).get_path(), autorestart, in_developer_mode=in_developer_mode)
-
-    if command in START_COMMANDS:
-        log.info('Agent version %s' % get_version())
 
     if 'start' == command:
         log.info('Start daemon')
@@ -420,6 +515,7 @@ def main():
         return Agent.info(verbose=options.verbose)
 
     elif 'foreground' == command:
+        log.info('Agent version %s' % get_version())
         if autorestart:
             # Set-up the supervisor callbacks and fork it.
             logging.info('Running Agent with auto-restart ON')