+++ checks.d/process.py
@@ -307,7 +307,15 @@ class ProcessCheck(AgentCheck):
         search_string = instance.get('search_string', None)
         ignore_ad = _is_affirmative(instance.get('ignore_denied_access', True))
         pid = instance.get('pid')
-        pid_file = instance.get('pid_file', None)
+        pid_file = instance.get('pid_file')
+
+        if self._conflicting_procfs:
+            self.warning('The `procfs_path` defined in `process.yaml` is different from the one defined in '
+                         '`datadog.conf`. This is currently not supported by the Agent. Defaulting to the '
+                         'value defined in `datadog.conf`: {}'.format(psutil.PROCFS_PATH))
+        elif self._deprecated_init_procfs:
+            self.warning('DEPRECATION NOTICE: Specifying `procfs_path` in `process.yaml` is deprecated. '
+                         'Please specify it in `datadog.conf` instead')
 
         if self._conflicting_procfs:
             self.warning('The `procfs_path` defined in `process.yaml` is different from the one defined in '
@@ -340,11 +348,11 @@ class ProcessCheck(AgentCheck):
         elif pid is not None:
             # we use Process(pid) as a means to search, if pid not found
             # psutil.NoSuchProcess is raised.
-            pids = set([psutil.Process(pid).pid])
+            pids = self._get_pid_set(pid)
         elif pid_file is not None:
             with open(pid_file, 'r') as file_pid:
                 pid_line = file_pid.readline().strip()
-                pids = set([psutil.Process(int(pid_line)).pid])
+                pids = self._get_pid_set(int(pid_line))
         else:
             raise ValueError('The "search_string" or "pid" options are required for process identification')
 
@@ -355,6 +363,9 @@ class ProcessCheck(AgentCheck):
 
         self.log.debug('ProcessCheck: process %s analysed', name)
         self.gauge('system.processes.number', len(pids), tags=tags)
+
+        if len(pids) == 0:
+            self.warning("No matching process was found")
 
         for attr, mname in ATTR_TO_METRIC.iteritems():
             vals = [x for x in proc_state[attr] if x is not None]
@@ -374,16 +385,23 @@ class ProcessCheck(AgentCheck):
             if vals:
                 self.rate('system.processes.%s' % mname, sum(vals), tags=tags)
 
-        self._process_service_check(name, len(pids), instance.get('thresholds', None))
+        self._process_service_check(name, len(pids), instance.get('thresholds', None), tags)
 
-    def _process_service_check(self, name, nb_procs, bounds):
-        '''
+    def _get_pid_set(self, pid):
+        try:
+            return {psutil.Process(pid).pid}
+        except psutil.NoSuchProcess:
+            return set()
+
+    def _process_service_check(self, name, nb_procs, bounds, tags):
+        """
         Report a service check, for each process in search_string.
         Report as OK if the process is in the warning thresholds
                    CRITICAL             out of the critical thresholds
                    WARNING              out of the warning thresholds
-        '''
-        tag = ["process:%s" % name]
+        """
+        # FIXME 6.x remove the `process:name` tag
+        service_check_tags = tags + ["process:%s" % name]
         status = AgentCheck.OK
         message_str = "PROCS %s: %s processes found for %s"
         status_str = {
@@ -406,6 +424,6 @@ class ProcessCheck(AgentCheck):
         self.service_check(
             "process.up",
             status,
-            tags=tag,
+            tags=service_check_tags,
             message=message_str % (status_str[status], nb_procs, name)
         )